+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [5.21.2 (unreleased)](https://github.com/kdeldycke/meta-package-manager/compare/v5.21.1...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [5.21.1 (2025-08-25)](https://github.com/kdeldycke/meta-package-manager/compare/v5.21.0...v5.21.1)
 
 - [scoop] Fix parsing of `scoop` version to support raw Git output.